requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
youtube-transcript-api>=0.6.2
pyyaml>=6.0
markdown>=3.5
reportlab>=4.1.0
